refactor: rename packages/notro to packages/notro-loader
- Rename packages/notro/ directory to packages/notro-loader/
- Change npm package name from "notro" to "notro-loader"
- Update all import paths in templates: "notro" -> "notro-loader",
  "notro/utils" -> "notro-loader/utils", "notro/integration" -> "notro-loader/integration"
- Update template package.json dependencies accordingly
- Update create-notro giget URL and repository URLs to mosugi/notro
- Update CLAUDE.md package name references and repository structure
- Update root package.json repository URL
