+++ templates/blog/astro.config.mjs
@@ -3,7 +3,7 @@ import sitemap from "@astrojs/sitemap";
 import partytown from "@astrojs/partytown";
 import tailwindcss from "@tailwindcss/vite";
 import { notionImageServiceConfig } from "./src/lib/notionImageService.js";
-import { notro } from "notro/integration";
+import { notro } from "notro-loader/integration";
 import { rehypeMermaid } from "rehype-beautiful-mermaid";
 import remarkMath from "remark-math";
 import rehypeKatex from "rehype-katex";
