+++ templates/blog/package.json
@@ -27,7 +27,7 @@
     "@astrojs/sitemap": "^3.7.1",
     "@tailwindcss/vite": "^4.2.1",
     "astro": "^6.0.4",
-    "notro": "^0.0.1",
+    "notro-loader": "^0.0.1",
     "tailwind-merge": "^3.5.0",
     "tailwind-variants": "^3.2.2",
     "tailwindcss": "^4.2.1"
