+++ templates/blog/src/components/blog/DatabaseCover.astro
@@ -1,6 +1,6 @@
 ---
 import { Image } from "astro:assets";
-import type { PageObjectResponseType } from "notro";
+import type { PageObjectResponseType } from "notro-loader";
 interface Props {
   cover: PageObjectResponseType["cover"];
   alt?: string;
