+++ templates/blog/src/components/blog/PostCard.astro
@@ -1,6 +1,6 @@
 ---
 import type { CollectionEntry } from "astro:content";
-import { getPlainText, getMultiSelect } from "notro";
+import { getPlainText, getMultiSelect } from "notro-loader";
 import DatabaseCover from "./DatabaseCover.astro";
 import config from "../../config";
