+++ templates/blog/src/pages/blog/[slug].astro
@@ -1,6 +1,6 @@
 ---
 import { getCollection } from "astro:content";
-import { buildLinkToPages, getPlainText, hasTag, NotroContent } from "notro";
+import { buildLinkToPages, getPlainText, hasTag, NotroContent } from "notro-loader";
 import Layout from "../../layouts/Layout.astro";
 import { notroComponents } from "../../components/notro";
 import config from "../../config";
