+++ templates/blog/src/pages/blog/tag/[tag]/[...page].astro
@@ -1,7 +1,7 @@
 ---
 import type { PaginateFunction } from "astro";
 import { getCollection } from "astro:content";
-import { hasTag } from "notro";
+import { hasTag } from "notro-loader";
 import BlogList from "../../../../components/blog/BlogList.astro";
 import Layout from "../../../../layouts/Layout.astro";
 import Pagination from "../../../../components/blog/Pagination.astro";
